requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
djangorestframework==3.13.1
idna==3.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
